Create README.md
Artemis Financial was the client, and they required a security assessment and refactoring of their existing software application to ensure secure communication and prevent data leakage.

During the security assessment, I identified and addressed the vulnerabilities in the codebase, including cryptography, code quality, client, and server. I used industry best practices to check and access the code frequently for any vulnerabilities, particularly after executing new functions or changing current ones before publishing it to the live code. This is important because coding securely adds value to a company's overall well-being by safeguarding its products and assets and ensuring trust is built with clients.

One challenging aspect of the vulnerability assessment was determining which mitigation techniques to use for each vulnerability identified. To increase the layers of security, I created a signed certificate and generated keys for the application that allowed a connection of 128-bit AES encryption.

To ensure the code and software application were functional and secure, I checked the codebase for any new vulnerabilities after refactoring. I used various tools such as random number generators and AES encryption to secure the communication between the client and the server.

Future employers might find my work on this assignment as evidence of my skills, knowledge, and experience in software security assessment and refactoring. They might also find the use of industry-standard best practices, the tools and resources I used such as random number generators, and AES encryption helpful in future assignments or tasks.